+++ encoderfile/src/builder/config.rs
@@ -1,6 +1,4 @@
 use crate::common::{Config as EmbeddedConfig, LuaLibs, ModelConfig, model_type::ModelType};
-use crate::runtime::TaskType;
-use crate::runtime::InputType;
 use anyhow::{Context, Result, bail};
 use schemars::JsonSchema;
 use std::string::String;
