+++ encoderfile/src/inference/sequence_classification.rs
@@ -1,5 +1,5 @@
 use crate::{
-    common::{ModelConfig, SequenceClassificationResult}, error::ApiError, runtime::ClassifierState, transforms::{Postprocessor, SequenceClassificationTransform}
+    common::{SequenceClassificationResult}, error::ApiError, runtime::ClassifierState, transforms::{Postprocessor, SequenceClassificationTransform}
 };
 use ndarray::{Array2, Axis, Ix2};
 use ndarray_stats::QuantileExt;
